Fixed year slider UI
I tried implementing the year slider but the filtering code for trying to get link years within the ranges of the slider values doesn't seem to be working. I checked the console to make sure that it is pulling the slider values, so something is wrong with the section for deleting links.